Реализация PriorityQueue и улучшение UnboundedQueue
- Хранение очередей: std::array<unique_ptr<IQueue>, 2> queues_ ([0] - High, [1] - Normal)
- Потокобезопасность очереди с приоритетом:
  * мьютекс + условная переменная
  * queues_ - безопасное чтение (неизменяем после конструктора)
- Метод push():
  * прямой доступ к queues_ по индексу
  * проверка queues_ на выход за границы массива и nullptr
  * notify_one()
- Метод pop():
  * блокирующий wait с извлечением задачи в локальную переменную
  * отработка сигнала о завершении работы без потери задач
- Метод shutdown(): установка active_ = false, notify_all()

- Улучшена потокобезопасность UnboundedQueue:
  * Добавлен counting_semaphore busy_slots_ для единообразия с BoundedQueue
  * try_pop() теперь проверяет наличие задач через try_acquire() без блокировки

- Подробные поясняющие комментарии

Проект компилируется без ошибок

+++ include/queue/queue.hpp
@@ -1,13 +1,13 @@
 #pragma once
-#include <functional>
-#include <optional>
+#include <functional>  // подключение стандартного функционального объекта (типа задачи)
+#include <optional>    // подключение словарного типа опционального результата
 
 namespace dispatcher::queue {
 
 // Структура конфигурации очереди задач
 struct QueueOptions {
-    bool bounded;                 // true - ограниченная, false - неограниченная очередь
-    std::optional<int> capacity;  // max размер очереди (только для bounded-очереди)
+    bool bounded;                    // true - ограниченная, false - неограниченная очередь
+    std::optional<size_t> capacity;  // max размер очереди (только для bounded-очереди)
 };
 
 // Интерфейсный класс очереди задач
